Add domino package
It's required by Functest to ease integrating all OPNFV projects and
their requirements.

It also synchronizes requirements with stable/ocata and renames
domino-cli to domino_cli to conform with common module names.

It defines scripts and console_scripts to rely on $PATH.
run_multimode.sh is adapted to correctly call client, cli and server.

It should be noted that run.sh can raise conflicts if it's already
provided by another OPNFV package and that run_on_remote_nodes.sh is
not fully updated. Otherwise it obliges all OPNFV installers to be
updated as well.

run.sh is unchanged and it's called by releng from local directory.

It may rise runtime issues if several files are not loaded via
pkg_resources.

Change-Id: Ibd9dd56f316609180aa432c6b7ee5a375d9af674
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>